feat: conditional formatting module (#88)
* conditional formatting module

* Add conditional formatting module to worksheet

* Refactor addConditionalFormatting to add in ConditionalFormattingModule
